Enable vboot for sandbox and improve debugging/format strings
Some functions of vboot are disabled for sandbox because sandbox could not
support them. This has changed, so remove the sandbox #ifdefs in the code.

Some printf() strings cause warnings on sandbox - mostly that uin64_t is
not 'long long' on a 64-bit machine. The existing format strings in U-Boot
do not seem to take account of this, so add casts to remove the warnings.

Also add a few more debug strings to make it easier to see what is happening
in the vboot flow.

BUG=chrome-os-partner:21115
BRANCH=pit
TEST=manual
crosfw -b sandbox -V

See there are no warnings.

+++ firmware/lib/vboot_api_kernel.c
@@ -87,10 +87,12 @@ uint32_t VbTryLoadKernel(VbCommonParams *cparams, LoadKernelParams *p,
 		if (512 != disk_info[i].bytes_per_lba ||
 		    32 > disk_info[i].lba_count ||
 		    get_info_flags != disk_info[i].flags) {
-			VBDEBUG(("  skipping: bytes_per_lba=%lld "
+			VBDEBUG(("  skipping: bytes_per_lba=%llu "
 				 "lba_count=%lld flags=0x%x\n",
-				 disk_info[i].bytes_per_lba,
-				 disk_info[i].lba_count,
+				 (unsigned long long)
+					disk_info[i].bytes_per_lba,
+				 (unsigned long long)
+					disk_info[i].lba_count,
 				 disk_info[i].flags));
 			continue;
 		}
