+++ firmware/lib/vboot_common_init.c
@@ -14,8 +14,8 @@
 
 int VbSharedDataInit(VbSharedDataHeader *header, uint64_t size)
 {
-	VBDEBUG(("VbSharedDataInit, %d bytes, header %d bytes\n", (int)size,
-		 sizeof(VbSharedDataHeader)));
+	VBDEBUG(("VbSharedDataInit, %d bytes, header %u bytes\n", (int)size,
+		 (unsigned)sizeof(VbSharedDataHeader)));
 
 	if (size < sizeof(VbSharedDataHeader)) {
 		VBDEBUG(("Not enough data for header.\n"));
